requirements.txt
# Interface
streamlit==1.45.1

# IA e NLP
openai==1.30.1
transformers==4.41.1
torch==2.3.0
nltk==3.8.1

# Voz
SpeechRecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.13  # Necessário para entrada de microfone

# Banco de dados
pandas==2.2.2
sqlite-utils==3.35  # Opcional para debugging de banco

# Integração com Google
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Utilitários
python-dotenv==1.0.1
requests==2.31.0
